components.converter renamed to components.exporter
