+++ src/datumaro/plugins/data_formats/camvid.py
@@ -20,9 +20,9 @@ from datumaro.components.annotation import (
     Mask,
     MaskCategories,
 )
-from datumaro.components.converter import Converter
 from datumaro.components.dataset import ItemStatus
 from datumaro.components.errors import MediaTypeError
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import DatasetItem, Importer, SourceExtractor
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image
