+++ src/datumaro/plugins/data_formats/cityscapes.py
@@ -21,9 +21,9 @@ from datumaro.components.annotation import (
     MaskCategories,
     RgbColor,
 )
-from datumaro.components.converter import Converter
 from datumaro.components.dataset import ItemStatus
 from datumaro.components.errors import MediaTypeError
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import CategoriesInfo, DatasetItem, Importer, SourceExtractor
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image
