+++ src/datumaro/plugins/data_formats/coco/exporter.py
@@ -20,9 +20,9 @@ from datumaro.components.annotation import (
     Points,
     Skeleton,
 )
-from datumaro.components.converter import Converter
 from datumaro.components.dataset import ItemStatus
 from datumaro.components.errors import MediaTypeError
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import DatasetItem
 from datumaro.components.media import Image
 from datumaro.util import cast, dump_json_file, find, str_to_bool
