+++ src/datumaro/plugins/data_formats/cvat/exporter.py
@@ -13,9 +13,9 @@ from itertools import chain
 from xml.sax.saxutils import XMLGenerator  # nosec
 
 from datumaro.components.annotation import AnnotationType, LabelCategories, PointsCategories
-from datumaro.components.converter import Converter
 from datumaro.components.dataset import ItemStatus
 from datumaro.components.errors import MediaTypeError
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import DatasetItem
 from datumaro.components.media import Image
 from datumaro.util import cast, pairs
