+++ src/datumaro/plugins/data_formats/image_dir.py
@@ -6,7 +6,7 @@ import logging as log
 import os
 import os.path as osp
 
-from datumaro.components.converter import Converter
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import DatasetItem, Importer, SourceExtractor
 from datumaro.components.media import Image
 from datumaro.util.image import find_images
