+++ src/datumaro/plugins/data_formats/image_zip.py
@@ -8,7 +8,7 @@ import os.path as osp
 from enum import Enum
 from zipfile import ZIP_BZIP2, ZIP_DEFLATED, ZIP_LZMA, ZIP_STORED, ZipFile
 
-from datumaro.components.converter import Converter
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import DatasetItem, Importer, SourceExtractor
 from datumaro.components.media import ByteImage
 from datumaro.util import parse_str_enum_value
