+++ src/datumaro/plugins/data_formats/kitti/exporter.py
@@ -12,8 +12,8 @@ from enum import Enum, auto
 import numpy as np
 
 from datumaro.components.annotation import AnnotationType, CompiledMask, LabelCategories
-from datumaro.components.converter import Converter
 from datumaro.components.errors import MediaTypeError
+from datumaro.components.exporter import Converter
 from datumaro.components.media import Image
 from datumaro.util import cast, parse_str_enum_value, str_to_bool
 from datumaro.util.annotation_util import make_label_id_mapping
