+++ src/datumaro/plugins/data_formats/lfw.py
@@ -7,8 +7,8 @@ import os.path as osp
 import re
 
 from datumaro.components.annotation import AnnotationType, Label, LabelCategories, Points
-from datumaro.components.converter import Converter
 from datumaro.components.errors import MediaTypeError
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import DatasetItem, Importer, SourceExtractor
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image
