+++ src/datumaro/plugins/data_formats/market1501.py
@@ -6,8 +6,8 @@ import os
 import os.path as osp
 import re
 
-from datumaro.components.converter import Converter
 from datumaro.components.errors import MediaTypeError
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import DatasetItem, Extractor, Importer
 from datumaro.components.media import Image
 from datumaro.util import str_to_bool
