+++ src/datumaro/plugins/data_formats/mot.py
@@ -15,8 +15,8 @@ from collections import OrderedDict
 from enum import Enum
 
 from datumaro.components.annotation import AnnotationType, Bbox, LabelCategories
-from datumaro.components.converter import Converter
 from datumaro.components.errors import MediaTypeError
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import DatasetItem, Importer, SourceExtractor
 from datumaro.components.format_detection import FormatDetectionContext
 from datumaro.components.media import Image
