+++ src/datumaro/plugins/data_formats/yolo/exporter.py
@@ -24,9 +24,9 @@ from datumaro.components.annotation import (
     Polygon,
     Skeleton,
 )
-from datumaro.components.converter import Converter
 from datumaro.components.dataset import DatasetPatch, ItemStatus
 from datumaro.components.errors import DatasetExportError, MediaTypeError
+from datumaro.components.exporter import Converter
 from datumaro.components.extractor import DEFAULT_SUBSET_NAME, DatasetItem, IExtractor
 from datumaro.components.media import Image
 from datumaro.util import dump_json_file, str_to_bool
